refactor(execution-evm): split lib logic into modules
