+++ crates/execution/evm/src/config.rs
@@ -1,4 +1,38 @@
-use revm::primitives::{Address, B256, Bytes};
+use alloc::sync::Arc;
+use core::{fmt::Debug, marker::PhantomData};
+
+use alloy_consensus::{BlockHeader, Header};
+#[cfg(feature = "std")]
+use alloy_eips::Decodable2718;
+use alloy_evm::{EvmFactory, FromRecoveredTx, FromTxWithEncoded};
+#[cfg(feature = "std")]
+use alloy_primitives::Bytes;
+use base_common_chains::Upgrades;
+use base_common_consensus::{BasePrimitives, DepositReceiptExt, EIP1559ParamError};
+use base_common_evm::{
+    BaseBlockExecutionCtx, BaseBlockExecutorFactory, BaseEvmFactory, BaseReceiptBuilder,
+    BaseSpecId, BaseTransaction, BaseTxEnv,
+};
+#[cfg(not(feature = "std"))]
+use base_common_rpc_types_engine as _;
+#[cfg(feature = "std")]
+use base_common_rpc_types_engine::ExecutionData;
+use base_execution_chainspec::BaseChainSpec;
+use reth_chainspec::EthChainSpec;
+#[cfg(feature = "std")]
+use reth_evm::{ConfigureEngineEvm, EvmEnvFor, ExecutableTxIterator, ExecutionCtxFor};
+use reth_evm::{ConfigureEvm, EvmEnv, TransactionEnv, precompiles::PrecompilesMap};
+use reth_primitives_traits::{NodePrimitives, SealedBlock, SealedHeader, SignedTransaction};
+#[cfg(feature = "std")]
+use reth_primitives_traits::{TxTy, WithEncoded};
+#[cfg(not(feature = "std"))]
+use reth_storage_errors as _;
+#[cfg(feature = "std")]
+use reth_storage_errors::any::AnyError;
+use revm::context::{BlockEnv, TxEnv};
+use revm::primitives::{Address, B256, Bytes as RevmBytes};
+
+use crate::{BaseBlockAssembler, BaseEvmEnvBuilder, BaseRethReceiptBuilder};
 
 /// Context relevant for execution of a next Base block.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -14,7 +48,7 @@ pub struct BaseNextBlockEnvAttributes {
     /// The parent beacon block root.
     pub parent_beacon_block_root: Option<B256>,
     /// Encoded EIP-1559 parameters to include into block's `extra_data` field.
-    pub extra_data: Bytes,
+    pub extra_data: RevmBytes,
 }
 
 #[cfg(feature = "rpc")]
@@ -30,5 +64,830 @@ impl<H: alloy_consensus::BlockHeader> reth_rpc_eth_api::helpers::pending_block::
             parent_beacon_block_root: parent.parent_beacon_block_root(),
             extra_data: parent.extra_data().clone(),
         }
+    }
+}
+
+/// Base EVM configuration.
+#[derive(Debug)]
+pub struct BaseEvmConfig<
+    ChainSpec = BaseChainSpec,
+    N: NodePrimitives = BasePrimitives,
+    R = BaseRethReceiptBuilder,
+    EvmFactory = BaseEvmFactory,
+> {
+    /// Inner [`BaseBlockExecutorFactory`].
+    pub executor_factory: BaseBlockExecutorFactory<R, Arc<ChainSpec>, EvmFactory>,
+    /// Base block assembler.
+    pub block_assembler: BaseBlockAssembler<ChainSpec>,
+    #[doc(hidden)]
+    pub _pd: PhantomData<N>,
+}
+
+impl<ChainSpec, N: NodePrimitives, R: Clone, EvmFactory: Clone> Clone
+    for BaseEvmConfig<ChainSpec, N, R, EvmFactory>
+{
+    fn clone(&self) -> Self {
+        Self {
+            executor_factory: self.executor_factory.clone(),
+            block_assembler: self.block_assembler.clone(),
+            _pd: self._pd,
+        }
+    }
+}
+
+impl<ChainSpec: Upgrades> BaseEvmConfig<ChainSpec> {
+    /// Creates a new [`BaseEvmConfig`] with the given chain spec for Base chains.
+    pub fn base(chain_spec: Arc<ChainSpec>) -> Self {
+        Self::new(chain_spec, BaseRethReceiptBuilder)
+    }
+}
+
+impl<ChainSpec: Upgrades, N: NodePrimitives, R> BaseEvmConfig<ChainSpec, N, R> {
+    /// Creates a new [`BaseEvmConfig`] with the given chain spec.
+    pub fn new(chain_spec: Arc<ChainSpec>, receipt_builder: R) -> Self {
+        Self {
+            block_assembler: BaseBlockAssembler::new(Arc::clone(&chain_spec)),
+            executor_factory: BaseBlockExecutorFactory::new(
+                receipt_builder,
+                chain_spec,
+                BaseEvmFactory::default(),
+            ),
+            _pd: PhantomData,
+        }
+    }
+}
+
+impl<ChainSpec, N, R, EvmFactory> BaseEvmConfig<ChainSpec, N, R, EvmFactory>
+where
+    ChainSpec: Upgrades,
+    N: NodePrimitives,
+{
+    /// Returns the chain spec associated with this configuration.
+    pub const fn chain_spec(&self) -> &Arc<ChainSpec> {
+        self.executor_factory.spec()
+    }
+}
+
+impl<ChainSpec, N, R, EvmF> ConfigureEvm for BaseEvmConfig<ChainSpec, N, R, EvmF>
+where
+    ChainSpec: EthChainSpec<Header = Header> + Upgrades,
+    N: NodePrimitives<
+            Receipt = R::Receipt,
+            SignedTx = R::Transaction,
+            BlockHeader = Header,
+            BlockBody = alloy_consensus::BlockBody<R::Transaction>,
+            Block = alloy_consensus::Block<R::Transaction>,
+        >,
+    BaseTransaction<TxEnv>: FromRecoveredTx<N::SignedTx> + FromTxWithEncoded<N::SignedTx>,
+    R: BaseReceiptBuilder<Receipt: DepositReceiptExt, Transaction: SignedTransaction>,
+    EvmF: EvmFactory<
+            Tx: FromRecoveredTx<R::Transaction>
+                    + FromTxWithEncoded<R::Transaction>
+                    + TransactionEnv
+                    + BaseTxEnv,
+            Precompiles = PrecompilesMap,
+            Spec = BaseSpecId,
+            BlockEnv = BlockEnv,
+        > + Debug,
+    Self: Send + Sync + Unpin + Clone + 'static,
+{
+    type Primitives = N;
+    type Error = EIP1559ParamError;
+    type NextBlockEnvCtx = BaseNextBlockEnvAttributes;
+    type BlockExecutorFactory = BaseBlockExecutorFactory<R, Arc<ChainSpec>, EvmF>;
+    type BlockAssembler = BaseBlockAssembler<ChainSpec>;
+
+    fn block_executor_factory(&self) -> &Self::BlockExecutorFactory {
+        &self.executor_factory
+    }
+
+    fn block_assembler(&self) -> &Self::BlockAssembler {
+        &self.block_assembler
+    }
+
+    fn evm_env(&self, header: &Header) -> Result<EvmEnv<BaseSpecId>, Self::Error> {
+        Ok(BaseEvmEnvBuilder::evm_env(header, self.chain_spec()))
+    }
+
+    fn next_evm_env(
+        &self,
+        parent: &Header,
+        attributes: &Self::NextBlockEnvCtx,
+    ) -> Result<EvmEnv<BaseSpecId>, Self::Error> {
+        let base_fee =
+            self.chain_spec().next_block_base_fee(parent, attributes.timestamp).unwrap_or_default();
+
+        Ok(BaseEvmEnvBuilder::next_evm_env(parent, attributes, base_fee, self.chain_spec()))
+    }
+
+    fn context_for_block(
+        &self,
+        block: &'_ SealedBlock<N::Block>,
+    ) -> Result<BaseBlockExecutionCtx, Self::Error> {
+        Ok(BaseBlockExecutionCtx {
+            parent_hash: block.header().parent_hash(),
+            parent_beacon_block_root: block.header().parent_beacon_block_root(),
+            extra_data: block.header().extra_data().clone(),
+        })
+    }
+
+    fn context_for_next_block(
+        &self,
+        parent: &SealedHeader<N::BlockHeader>,
+        attributes: Self::NextBlockEnvCtx,
+    ) -> Result<BaseBlockExecutionCtx, Self::Error> {
+        Ok(BaseBlockExecutionCtx {
+            parent_hash: parent.hash(),
+            parent_beacon_block_root: attributes.parent_beacon_block_root,
+            extra_data: attributes.extra_data,
+        })
+    }
+}
+
+#[cfg(feature = "std")]
+impl<ChainSpec, N, R> ConfigureEngineEvm<ExecutionData> for BaseEvmConfig<ChainSpec, N, R>
+where
+    ChainSpec: EthChainSpec<Header = Header> + Upgrades,
+    N: NodePrimitives<
+            Receipt = R::Receipt,
+            SignedTx = R::Transaction,
+            BlockHeader = Header,
+            BlockBody = alloy_consensus::BlockBody<R::Transaction>,
+            Block = alloy_consensus::Block<R::Transaction>,
+        >,
+    BaseTransaction<TxEnv>: FromRecoveredTx<N::SignedTx> + FromTxWithEncoded<N::SignedTx>,
+    R: BaseReceiptBuilder<Receipt: DepositReceiptExt, Transaction: SignedTransaction>,
+    Self: Send + Sync + Unpin + Clone + 'static,
+{
+    fn evm_env_for_payload(&self, payload: &ExecutionData) -> Result<EvmEnvFor<Self>, Self::Error> {
+        Ok(BaseEvmEnvBuilder::payload_evm_env(payload, self.chain_spec()))
+    }
+
+    fn context_for_payload<'a>(
+        &self,
+        payload: &'a ExecutionData,
+    ) -> Result<ExecutionCtxFor<'a, Self>, Self::Error> {
+        Ok(BaseBlockExecutionCtx {
+            parent_hash: payload.parent_hash(),
+            parent_beacon_block_root: payload.sidecar.parent_beacon_block_root(),
+            extra_data: payload.payload.as_v1().extra_data.clone(),
+        })
+    }
+
+    fn tx_iterator_for_payload(
+        &self,
+        payload: &ExecutionData,
+    ) -> Result<impl ExecutableTxIterator<Self>, Self::Error> {
+        let transactions = payload.payload.transactions().clone();
+        let convert = |encoded: Bytes| {
+            let tx = TxTy::<Self::Primitives>::decode_2718_exact(encoded.as_ref())
+                .map_err(AnyError::new)?;
+            let signer = tx.try_recover().map_err(AnyError::new)?;
+            Ok::<_, AnyError>(WithEncoded::new(encoded, tx.with_signer(signer)))
+        };
+
+        Ok((transactions, convert))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use alloc::collections::BTreeMap;
+    use std::sync::Arc;
+
+    use alloy_consensus::{Header, Receipt};
+    use alloy_eips::eip7685::Requests;
+    use alloy_genesis::Genesis;
+    use alloy_primitives::{
+        Address, B256, LogData, U256, bytes,
+        map::{AddressMap, B256Map, HashMap},
+    };
+    use base_common_chains::BaseUpgrade;
+    use base_common_consensus::{BaseBlock, BasePrimitives, BaseReceipt};
+    use base_common_evm::BaseSpecId;
+    use base_execution_chainspec::{BaseChainSpec, BaseChainSpecBuilder};
+    use reth_chainspec::ChainSpec;
+    use reth_evm::{ConfigureEvm, EvmEnv, execute::ProviderError};
+    use reth_execution_types::{
+        AccountRevertInit, BundleStateInit, Chain, ExecutionOutcome, RevertsInit,
+    };
+    use reth_primitives_traits::{Account, RecoveredBlock, constants::MAX_TX_GAS_LIMIT_OSAKA};
+    use revm::{
+        context::{BlockEnv, CfgEnv},
+        database::{BundleState, CacheDB},
+        database_interface::EmptyDBTyped,
+        inspector::NoOpInspector,
+        primitives::Log,
+        state::AccountInfo,
+    };
+
+    use super::BaseEvmConfig;
+
+    fn test_evm_config() -> BaseEvmConfig {
+        BaseEvmConfig::base(Arc::new(BaseChainSpec::mainnet()))
+    }
+
+    #[test]
+    fn test_evm_env_uses_azul_for_genesis_chain_spec() {
+        let chain_spec = Arc::new(
+            BaseChainSpecBuilder::default()
+                .chain(0.into())
+                .genesis(Genesis::default())
+                .azul_activated()
+                .build(),
+        );
+        let evm_config = BaseEvmConfig::base(chain_spec);
+        let header = Header { timestamp: 0, ..Default::default() };
+        let EvmEnv { cfg_env, .. } = evm_config.evm_env(&header).unwrap();
+        assert_eq!(cfg_env.spec, BaseSpecId::new(BaseUpgrade::Azul));
+        assert_eq!(cfg_env.tx_gas_limit_cap, Some(MAX_TX_GAS_LIMIT_OSAKA));
+    }
+
+    #[test]
+    fn test_fill_cfg_and_block_env() {
+        // Create a default header
+        let header = Header::default();
+
+        // Build the ChainSpec for Ethereum mainnet, activating London, Paris, and Shanghai
+        // hardforks
+        let chain_spec = ChainSpec::builder()
+            .chain(0.into())
+            .genesis(Genesis::default())
+            .london_activated()
+            .paris_activated()
+            .shanghai_activated()
+            .build();
+
+        // Use the `BaseEvmConfig` to create the `cfg_env` and `block_env` based on the ChainSpec,
+        // Header, and total difficulty
+        let EvmEnv { cfg_env, .. } =
+            BaseEvmConfig::base(Arc::new(BaseChainSpec { inner: chain_spec.clone() }))
+                .evm_env(&header)
+                .unwrap();
+
+        // Assert that the chain ID in the `cfg_env` is correctly set to the chain ID of the
+        // ChainSpec
+        assert_eq!(cfg_env.chain_id, chain_spec.chain().id());
+    }
+
+    #[test]
+    fn test_evm_with_env_default_spec() {
+        let evm_config = test_evm_config();
+
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        let evm_env = EvmEnv::default();
+
+        let evm = evm_config.evm_with_env(db, evm_env.clone());
+
+        // Check that the EVM environment
+        assert_eq!(evm.cfg, evm_env.cfg_env);
+    }
+
+    #[test]
+    fn test_evm_with_env_custom_cfg() {
+        let evm_config = test_evm_config();
+
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        // Create a custom configuration environment with a chain ID of 111
+        let cfg = CfgEnv::new()
+            .with_chain_id(111)
+            .with_spec_and_mainnet_gas_params(BaseSpecId::default());
+
+        let evm_env = EvmEnv { cfg_env: cfg.clone(), ..Default::default() };
+
+        let evm = evm_config.evm_with_env(db, evm_env);
+
+        // Check that the EVM environment is initialized with the custom environment
+        assert_eq!(evm.cfg, cfg);
+    }
+
+    #[test]
+    fn test_evm_with_env_custom_block_and_tx() {
+        let evm_config = test_evm_config();
+
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        // Create customs block and tx env
+        let block = BlockEnv {
+            basefee: 1000,
+            gas_limit: 10_000_000,
+            number: U256::from(42),
+            ..Default::default()
+        };
+
+        let evm_env = EvmEnv { block_env: block, ..Default::default() };
+
+        let evm = evm_config.evm_with_env(db, evm_env.clone());
+
+        // Verify that the block and transaction environments are set correctly
+        assert_eq!(evm.block, evm_env.block_env);
+    }
+
+    #[test]
+    fn test_evm_with_spec_id() {
+        let evm_config = test_evm_config();
+
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        let evm_env = EvmEnv {
+            cfg_env: CfgEnv::new()
+                .with_spec_and_mainnet_gas_params(BaseSpecId::new(BaseUpgrade::Ecotone)),
+            ..Default::default()
+        };
+
+        let evm = evm_config.evm_with_env(db, evm_env.clone());
+
+        assert_eq!(evm.cfg, evm_env.cfg_env);
+    }
+
+    #[test]
+    fn test_evm_with_env_and_default_inspector() {
+        let evm_config = test_evm_config();
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        let evm_env = EvmEnv { cfg_env: Default::default(), ..Default::default() };
+
+        let evm = evm_config.evm_with_env_and_inspector(db, evm_env.clone(), NoOpInspector {});
+
+        // Check that the EVM environment is set to default values
+        assert_eq!(evm.block, evm_env.block_env);
+        assert_eq!(evm.cfg, evm_env.cfg_env);
+    }
+
+    #[test]
+    fn test_evm_with_env_inspector_and_custom_cfg() {
+        let evm_config = test_evm_config();
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        let cfg = CfgEnv::new()
+            .with_chain_id(111)
+            .with_spec_and_mainnet_gas_params(BaseSpecId::default());
+        let block = BlockEnv::default();
+        let evm_env = EvmEnv { block_env: block, cfg_env: cfg.clone() };
+
+        let evm = evm_config.evm_with_env_and_inspector(db, evm_env.clone(), NoOpInspector {});
+
+        // Check that the EVM environment is set with custom configuration
+        assert_eq!(evm.cfg, cfg);
+        assert_eq!(evm.block, evm_env.block_env);
+    }
+
+    #[test]
+    fn test_evm_with_env_inspector_and_custom_block_tx() {
+        let evm_config = test_evm_config();
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        // Create custom block and tx environment
+        let block = BlockEnv {
+            basefee: 1000,
+            gas_limit: 10_000_000,
+            number: U256::from(42),
+            ..Default::default()
+        };
+        let evm_env = EvmEnv { block_env: block, ..Default::default() };
+
+        let evm = evm_config.evm_with_env_and_inspector(db, evm_env.clone(), NoOpInspector {});
+
+        // Verify that the block and transaction environments are set correctly
+        assert_eq!(evm.block, evm_env.block_env);
+    }
+
+    #[test]
+    fn test_evm_with_env_inspector_and_spec_id() {
+        let evm_config = test_evm_config();
+        let db = CacheDB::<EmptyDBTyped<ProviderError>>::default();
+
+        let evm_env = EvmEnv {
+            cfg_env: CfgEnv::new()
+                .with_spec_and_mainnet_gas_params(BaseSpecId::new(BaseUpgrade::Ecotone)),
+            ..Default::default()
+        };
+
+        let evm = evm_config.evm_with_env_and_inspector(db, evm_env.clone(), NoOpInspector {});
+
+        // Check that the spec ID is set properly
+        assert_eq!(evm.cfg, evm_env.cfg_env);
+        assert_eq!(evm.block, evm_env.block_env);
+    }
+
+    #[test]
+    fn receipts_by_block_hash() {
+        // Create a default recovered block
+        let block: RecoveredBlock<BaseBlock> = Default::default();
+
+        // Define block hashes for block1 and block2
+        let block1_hash = B256::new([0x01; 32]);
+        let block2_hash = B256::new([0x02; 32]);
+
+        // Clone the default block into block1 and block2
+        let mut block1 = block.clone();
+        let mut block2 = block;
+
+        // Set the hashes of block1 and block2
+        block1.set_block_number(10);
+        block1.set_hash(block1_hash);
+
+        block2.set_block_number(11);
+        block2.set_hash(block2_hash);
+
+        // Create a random receipt object, receipt1
+        let receipt1 = BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![],
+            status: true.into(),
+        });
+
+        // Create another random receipt object, receipt2
+        let receipt2 = BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 1325345,
+            logs: vec![],
+            status: true.into(),
+        });
+
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![receipt1.clone()], vec![receipt2]];
+
+        // Create an ExecutionOutcome object with the created bundle, receipts, an empty requests
+        // vector, and first_block set to 10
+        let execution_outcome = ExecutionOutcome::<BaseReceipt> {
+            bundle: Default::default(),
+            receipts,
+            requests: vec![],
+            first_block: 10,
+        };
+
+        // Create a Chain object with a BTreeMap of blocks mapped to their block numbers,
+        // including block1_hash and block2_hash, and the execution_outcome
+        let chain: Chain<BasePrimitives> =
+            Chain::new([block1, block2], execution_outcome.clone(), BTreeMap::new());
+
+        // Assert that the proper receipt vector is returned for block1_hash
+        assert_eq!(chain.receipts_by_block_hash(block1_hash), Some(vec![&receipt1]));
+
+        // Create an ExecutionOutcome object with a single receipt vector containing receipt1
+        let execution_outcome1 = ExecutionOutcome {
+            bundle: Default::default(),
+            receipts: vec![vec![receipt1]],
+            requests: vec![],
+            first_block: 10,
+        };
+
+        // Assert that the execution outcome at the first block contains only the first receipt
+        assert_eq!(chain.execution_outcome_at_block(10), Some(execution_outcome1));
+
+        // Assert that the execution outcome at the tip block contains the whole execution outcome
+        assert_eq!(chain.execution_outcome_at_block(11), Some(execution_outcome));
+    }
+
+    #[test]
+    fn test_initialization() {
+        // Create a new BundleState object with initial data
+        let bundle = BundleState::new(
+            vec![(Address::new([2; 20]), None, Some(AccountInfo::default()), HashMap::default())],
+            vec![vec![(Address::new([2; 20]), None, vec![])]],
+            vec![],
+        );
+
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![Some(BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![],
+            status: true.into(),
+        }))]];
+
+        // Create a Requests object with a vector of requests
+        let requests = vec![Requests::new(vec![bytes!("dead"), bytes!("beef"), bytes!("beebee")])];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let exec_res = ExecutionOutcome {
+            bundle: bundle.clone(),
+            receipts: receipts.clone(),
+            requests: requests.clone(),
+            first_block,
+        };
+
+        // Assert that creating a new ExecutionOutcome using the constructor matches exec_res
+        assert_eq!(
+            ExecutionOutcome::new(bundle, receipts.clone(), first_block, requests.clone()),
+            exec_res
+        );
+
+        // Create a BundleStateInit object and insert initial data
+        let mut state_init: BundleStateInit = AddressMap::default();
+        state_init
+            .insert(Address::new([2; 20]), (None, Some(Account::default()), B256Map::default()));
+
+        // Create an AddressMap for account reverts and insert initial data
+        let mut revert_inner: AddressMap<AccountRevertInit> = AddressMap::default();
+        revert_inner.insert(Address::new([2; 20]), (None, vec![]));
+
+        // Create a RevertsInit object and insert the revert_inner data
+        let mut revert_init: RevertsInit = HashMap::default();
+        revert_init.insert(123, revert_inner);
+
+        // Assert that creating a new ExecutionOutcome using the new_init method matches
+        // exec_res
+        assert_eq!(
+            ExecutionOutcome::new_init(
+                state_init,
+                revert_init,
+                vec![],
+                receipts,
+                first_block,
+                requests,
+            ),
+            exec_res
+        );
+    }
+
+    #[test]
+    fn test_block_number_to_index() {
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![Some(BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![],
+            status: true.into(),
+        }))]];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let exec_res = ExecutionOutcome {
+            bundle: Default::default(),
+            receipts,
+            requests: vec![],
+            first_block,
+        };
+
+        // Test before the first block
+        assert_eq!(exec_res.block_number_to_index(12), None);
+
+        // Test after the first block but index larger than receipts length
+        assert_eq!(exec_res.block_number_to_index(133), None);
+
+        // Test after the first block
+        assert_eq!(exec_res.block_number_to_index(123), Some(0));
+    }
+
+    #[test]
+    fn test_get_logs() {
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![Log::<LogData>::default()],
+            status: true.into(),
+        })]];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let exec_res = ExecutionOutcome {
+            bundle: Default::default(),
+            receipts,
+            requests: vec![],
+            first_block,
+        };
+
+        // Get logs for block number 123
+        let logs: Vec<&Log> = exec_res.logs(123).unwrap().collect();
+
+        // Assert that the logs match the expected logs
+        assert_eq!(logs, vec![&Log::<LogData>::default()]);
+    }
+
+    #[test]
+    fn test_receipts_by_block() {
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![Some(BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![Log::<LogData>::default()],
+            status: true.into(),
+        }))]];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let exec_res = ExecutionOutcome {
+            bundle: Default::default(), // Default value for bundle
+            receipts,                   // Include the created receipts
+            requests: vec![],           // Empty vector for requests
+            first_block,                // Set the first block number
+        };
+
+        // Get receipts for block number 123 and convert the result into a vector
+        let receipts_by_block: Vec<_> = exec_res.receipts_by_block(123).iter().collect();
+
+        // Assert that the receipts for block number 123 match the expected receipts
+        assert_eq!(
+            receipts_by_block,
+            vec![&Some(BaseReceipt::Legacy(Receipt::<Log> {
+                cumulative_gas_used: 46913,
+                logs: vec![Log::<LogData>::default()],
+                status: true.into(),
+            }))]
+        );
+    }
+
+    #[test]
+    fn test_receipts_len() {
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![Some(BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![Log::<LogData>::default()],
+            status: true.into(),
+        }))]];
+
+        // Create an empty Receipts object
+        let receipts_empty = vec![];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let exec_res = ExecutionOutcome {
+            bundle: Default::default(), // Default value for bundle
+            receipts,                   // Include the created receipts
+            requests: vec![],           // Empty vector for requests
+            first_block,                // Set the first block number
+        };
+
+        // Assert that the length of receipts in exec_res is 1
+        assert_eq!(exec_res.len(), 1);
+
+        // Assert that exec_res is not empty
+        assert!(!exec_res.is_empty());
+
+        // Create a ExecutionOutcome object with an empty Receipts object
+        let exec_res_empty_receipts: ExecutionOutcome<BaseReceipt> = ExecutionOutcome {
+            bundle: Default::default(), // Default value for bundle
+            receipts: receipts_empty,   // Include the empty receipts
+            requests: vec![],           // Empty vector for requests
+            first_block,                // Set the first block number
+        };
+
+        // Assert that the length of receipts in exec_res_empty_receipts is 0
+        assert_eq!(exec_res_empty_receipts.len(), 0);
+
+        // Assert that exec_res_empty_receipts is empty
+        assert!(exec_res_empty_receipts.is_empty());
+    }
+
+    #[test]
+    fn test_revert_to() {
+        // Create a random receipt object
+        let receipt = BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![],
+            status: true.into(),
+        });
+
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![vec![Some(receipt.clone())], vec![Some(receipt.clone())]];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a request.
+        let request = bytes!("deadbeef");
+
+        // Create a vector of Requests containing the request.
+        let requests =
+            vec![Requests::new(vec![request.clone()]), Requests::new(vec![request.clone()])];
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let mut exec_res =
+            ExecutionOutcome { bundle: Default::default(), receipts, requests, first_block };
+
+        // Assert that the revert_to method returns true when reverting to the initial block number.
+        assert!(exec_res.revert_to(123));
+
+        // Assert that the receipts are properly cut after reverting to the initial block number.
+        assert_eq!(exec_res.receipts, vec![vec![Some(receipt)]]);
+
+        // Assert that the requests are properly cut after reverting to the initial block number.
+        assert_eq!(exec_res.requests, vec![Requests::new(vec![request])]);
+
+        // Assert that the revert_to method returns false when attempting to revert to a block
+        // number greater than the initial block number.
+        assert!(!exec_res.revert_to(133));
+
+        // Assert that the revert_to method returns false when attempting to revert to a block
+        // number less than the initial block number.
+        assert!(!exec_res.revert_to(10));
+    }
+
+    #[test]
+    fn test_extend_execution_outcome() {
+        // Create a Receipt object with specific attributes.
+        let receipt = BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![],
+            status: true.into(),
+        });
+
+        // Create a Receipts object containing the receipt.
+        let receipts = vec![vec![Some(receipt.clone())]];
+
+        // Create a request.
+        let request = bytes!("deadbeef");
+
+        // Create a vector of Requests containing the request.
+        let requests = vec![Requests::new(vec![request.clone()])];
+
+        // Define the initial block number.
+        let first_block = 123;
+
+        // Create an ExecutionOutcome object.
+        let mut exec_res =
+            ExecutionOutcome { bundle: Default::default(), receipts, requests, first_block };
+
+        // Extend the ExecutionOutcome object by itself.
+        exec_res.extend(exec_res.clone());
+
+        // Assert the extended ExecutionOutcome matches the expected outcome.
+        assert_eq!(
+            exec_res,
+            ExecutionOutcome {
+                bundle: Default::default(),
+                receipts: vec![vec![Some(receipt.clone())], vec![Some(receipt)]],
+                requests: vec![Requests::new(vec![request.clone()]), Requests::new(vec![request])],
+                first_block: 123,
+            }
+        );
+    }
+
+    #[test]
+    fn test_split_at_execution_outcome() {
+        // Create a random receipt object
+        let receipt = BaseReceipt::Legacy(Receipt::<Log> {
+            cumulative_gas_used: 46913,
+            logs: vec![],
+            status: true.into(),
+        });
+
+        // Create a Receipts object with a vector of receipt vectors
+        let receipts = vec![
+            vec![Some(receipt.clone())],
+            vec![Some(receipt.clone())],
+            vec![Some(receipt.clone())],
+        ];
+
+        // Define the first block number
+        let first_block = 123;
+
+        // Create a request.
+        let request = bytes!("deadbeef");
+
+        // Create a vector of Requests containing the request.
+        let requests = vec![
+            Requests::new(vec![request.clone()]),
+            Requests::new(vec![request.clone()]),
+            Requests::new(vec![request.clone()]),
+        ];
+
+        // Create a ExecutionOutcome object with the created bundle, receipts, requests, and
+        // first_block
+        let exec_res =
+            ExecutionOutcome { bundle: Default::default(), receipts, requests, first_block };
+
+        // Split the ExecutionOutcome at block number 124
+        let result = exec_res.clone().split_at(124);
+
+        // Define the expected lower ExecutionOutcome after splitting
+        let lower_execution_outcome = ExecutionOutcome {
+            bundle: Default::default(),
+            receipts: vec![vec![Some(receipt.clone())]],
+            requests: vec![Requests::new(vec![request.clone()])],
+            first_block,
+        };
+
+        // Define the expected higher ExecutionOutcome after splitting
+        let higher_execution_outcome = ExecutionOutcome {
+            bundle: Default::default(),
+            receipts: vec![vec![Some(receipt.clone())], vec![Some(receipt)]],
+            requests: vec![Requests::new(vec![request.clone()]), Requests::new(vec![request])],
+            first_block: 124,
+        };
+
+        // Assert that the split result matches the expected lower and higher outcomes
+        assert_eq!(result.0, Some(lower_execution_outcome));
+        assert_eq!(result.1, higher_execution_outcome);
+
+        // Assert that splitting at the first block number returns None for the lower outcome
+        assert_eq!(exec_res.clone().split_at(123), (None, exec_res));
     }
 }
